@@ -4,7 +4,7 @@ import MongoABI
 extension Mongo
 {
     @frozen public
-    struct SortDocument:Mongo.EncodableDocument, Sendable
+    struct SortDocument:Sendable
     {
         public
         var bson:BSON.Document
@@ -16,95 +16,8 @@ extension Mongo
         }
     }
 }
-extension Mongo.SortDocument
+extension Mongo.SortDocument:Mongo.EncodableDocument
 {
-    @available(*, unavailable,
-        message: "pass the `(+)` or `(-)` operator functions to specify sort direction.")
-    @inlinable public
-    subscript(path:Mongo.AnyKeyPath) -> Int?
-    {
-        get
-        {
-            nil
-        }
-        set
-        {
-        }
-    }
-}
-extension Mongo.SortDocument
-{
-    @inlinable public
-    subscript(natural:Natural) -> ((Mongo.SortAscending) -> Never)?
-    {
-        get
-        {
-            nil
-        }
-        set(value)
-        {
-            if  case _? = value
-            {
-                (1 as Int32).encode(to: &self.bson[with: natural])
-            }
-        }
-    }
-    @inlinable public
-    subscript(path:Mongo.AnyKeyPath) -> ((Mongo.SortAscending) -> Never)?
-    {
-        get
-        {
-            nil
-        }
-        set(value)
-        {
-            if  case _? = value
-            {
-                (1 as Int32).encode(to: &self.bson[with: path.stem])
-            }
-        }
-    }
-
-    @inlinable public
-    subscript(natural:Natural) -> ((Mongo.SortDescending) -> Never)?
-    {
-        get
-        {
-            nil
-        }
-        set(value)
-        {
-            if  case _? = value
-            {
-                (-1 as Int32).encode(to: &self.bson[with: natural])
-            }
-        }
-    }
-    @inlinable public
-    subscript(path:Mongo.AnyKeyPath) -> ((Mongo.SortDescending) -> Never)?
-    {
-        get
-        {
-            nil
-        }
-        set(value)
-        {
-            if  case _? = value
-            {
-                (-1 as Int32).encode(to: &self.bson[with: path.stem])
-            }
-        }
-    }
-    @inlinable public
-    subscript(path:Mongo.AnyKeyPath) -> Mongo.SortOperator?
-    {
-        get
-        {
-            nil
-        }
-        set(value)
-        {
-            value?.encode(to: &self.bson[with: path.stem])
-        }
-    }
+    public
+    typealias Encoder = Mongo.SortEncoder
 }
