@@ -1,13 +1,18 @@
 extension Mongo
 {
-    public
+    @frozen public
     enum SortAscending
     {
     }
 }
+extension Mongo.SortAscending:Mongo.SortDirection
+{
+    @inlinable public static
+    var code:Int32 { 1 }
+}
 extension Mongo.SortAscending
 {
-    public static prefix
+    @inlinable public static prefix
     func + (lhs:Self) -> Never
     {
     }
